requirement.txt
Django==2.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
python-decouple==3.3
python-memcached==1.59